platform/registry-api/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
alembic==1.13.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Cloud SQL Python Connector (for IAM auth with private IP)
cloud-sql-python-connector[pg8000]==1.11.0
pg8000==1.31.2

# Google Auth (for ID token verification)
google-auth>=2.28.0
google-auth-oauthlib==1.2.0

# Additional utilities
python-multipart==0.0.6
requests==2.31.0
